celery-worker/requirements.txt
celery==5.3.4
redis==5.0.1
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
kombu==5.3.4